refactor: clean up and split qr_scanner
