+++ lib/screens/order_qr_scan.dart
@@ -3,7 +3,7 @@ import 'dart:developer';
 import 'package:beamer/beamer.dart';
 import 'package:flutter/material.dart';
 
-import '../widgets/qr_scanner.dart';
+import '../widgets/qr_scan/scanner.dart';
 
 class OrderQrScanScreen extends StatelessWidget {
   const OrderQrScanScreen({Key? key}) : super(key: key);
@@ -11,20 +11,18 @@ class OrderQrScanScreen extends StatelessWidget {
   static final orderCodePattern = RegExp(r'canteen-mgmt-order:(\d+)$');
 
   @override
-  Widget build(BuildContext context) => Scaffold(
-        body: QrScanner(
-          ignoreEmpty: true,
-          onScan: (context, scanData) {
-            if (scanData == null) return;
+  Widget build(BuildContext context) => QrScanner(
+        ignoreEmpty: true,
+        onScan: (context, scanData) {
+          if (scanData == null) return;
 
-            log(scanData);
+          log(scanData);
 
-            final match = orderCodePattern.matchAsPrefix(scanData);
-            if (match == null) return;
+          final match = orderCodePattern.matchAsPrefix(scanData);
+          if (match == null) return;
 
-            final orderId = match.group(1)!;
-            context.beamToNamed('/order/$orderId');
-          },
-        ),
+          final orderId = match.group(1)!;
+          context.beamToNamed('/order/$orderId');
+        },
       );
 }
